+++ src/types.ts
@@ -39,9 +39,13 @@ export type { Asset, Event, EventController, EventHandler, FontData, Vec2 };
  *
  * @group Start
  */
-declare function kaplay<T extends PluginList<unknown> = [undefined]>(
-    options?: KaboomOpt<T>,
-): T extends [undefined] ? KaboomCtx : KaboomCtx & MergePlugins<T>;
+declare function kaplay<
+    TPlugins extends PluginList<unknown> = [undefined],
+    TButtons extends ButtonsDef = {},
+>(
+    options?: KaboomOpt<TPlugins, TButtons>,
+): TPlugins extends [undefined] ? KaboomCtx<TButtons>
+    : KaboomCtx<TButtons> & MergePlugins<TPlugins>;
 
 export type InternalCtx = {
     kaboomCtx: KaboomCtx;
@@ -70,7 +74,7 @@ export type InternalCtx = {
  *
  * @group Start
  */
-export interface KaboomCtx {
+export interface KaboomCtx<TButtonDef extends ButtonsDef = {}> {
     /**
      * The internal context object.
      *
@@ -1310,6 +1314,27 @@ export interface KaboomCtx {
         action: (value: Vec2) => void,
     ): EventController;
     /**
+     * Register an event that runs when user press a defined button
+     * (like "jump") on any input (keyboard, gamepad).
+     */
+    onButtonPress(
+        button: keyof TButtonDef,
+        action: () => void,
+    ): EventController;
+    /**
+     * Register an event that runs when user release a defined button
+     * (like "jump") on any input (keyboard, gamepad).
+     */
+    onButtonRelease(
+        button: keyof TButtonDef,
+        action: () => void,
+    ): EventController;
+    /**
+     * Register an event that runs when user press a defined button
+     * (like "jump") on any input (keyboard, gamepad).
+     */
+    onButtonDown(button: keyof TButtonDef, action: () => void): EventController;
+    /**
      * Register an event that runs when current scene ends.
      *
      * @since v3000.0
@@ -1822,6 +1847,35 @@ export interface KaboomCtx {
      * @group Info
      */
     isGamepadButtonReleased(btn?: GamepadButton): boolean;
+    /**
+     * If a defined button is just pressed last frame on any input (keyboard, gamepad).
+     *
+     * @since v3001.0
+     * @group Info
+     */
+    isButtonPressed(button: keyof TButtonDef): boolean;
+    /**
+     * If a defined button is currently held down on any input (keyboard, gamepad).
+     *
+     * @since v3001.0
+     * @group Info
+     */
+    isButtonDown(button: keyof TButtonDef): boolean;
+    /**
+     * If a defined button is just released last frame on any input (keyboard, gamepad).
+     *
+     * @since v3001.0
+     * @group Info
+     */
+    isButtonReleased(button: keyof TButtonDef): boolean;
+    /**
+     * Get a input binding from a button name.
+     */
+    getButton(button: keyof TButtonDef): ButtonBinding;
+    /**
+     * Set a input binding for a button name.
+     */
+    setButton(button: string, def: ButtonBinding): void;
     /**
      * Get stick axis values from a gamepad.
      *
@@ -3369,6 +3423,22 @@ export type GamepadDef = {
     sticks: Partial<Record<GamepadStick, { x: number; y: number }>>;
 };
 
+/**
+ * A button binding.
+ */
+export type ButtonBinding = {
+    keyboard?: Key | Key[];
+    gamepad?: GamepadButton | GamepadButton[];
+    mouse?: MouseButton | MouseButton[];
+};
+
+/**
+ * A buttons definition.
+ */
+export type ButtonsDef = {
+    [key: string]: ButtonBinding;
+};
+
 /** A KAPLAY's gamepad */
 export type KGamePad = {
     /** The order of the gamepad in the gamepad list. */
@@ -3399,7 +3469,10 @@ export type GameObjInspect = Record<Tag, string | null>;
  *
  * @group Start
  */
-export interface KaboomOpt<T extends PluginList<any> = any> {
+export interface KaboomOpt<
+    TPlugin extends PluginList<any> = any,
+    TButtonDef extends ButtonsDef = any,
+> {
     /**
      * Width of game.
      */
@@ -3497,6 +3570,12 @@ export interface KaboomOpt<T extends PluginList<any> = any> {
      */
     gamepads?: Record<string, GamepadDef>;
     /**
+     * Defined buttons for input binding.
+     *
+     * @since v30010
+     */
+    buttons?: TButtonDef;
+    /**
      * Limit framerate to an amount per second.
      *
      * @since v3000.0
@@ -3515,7 +3594,7 @@ export interface KaboomOpt<T extends PluginList<any> = any> {
     /**
      * List of plugins to import.
      */
-    plugins?: T;
+    plugins?: TPlugin;
     /**
      * Enter burp mode.
      */